Rename DaemonThreadFactory to AgentThreadFactory
