+++ internal-api/src/main/java/datadog/trace/util/AgentThreadFactory.java
@@ -2,51 +2,50 @@ package datadog.trace.util;
 
 import java.util.concurrent.ThreadFactory;
 
-/** A {@link ThreadFactory} implementation that starts all {@link Thread} as daemons. */
-public final class DaemonThreadFactory implements ThreadFactory {
+/** A {@link ThreadFactory} implementation that starts all agent {@link Thread}s as daemons. */
+public final class AgentThreadFactory implements ThreadFactory {
   public static final ThreadGroup AGENT_THREAD_GROUP = new ThreadGroup("dd-trace-java");
 
-  public static final DaemonThreadFactory TRACE_MONITOR =
-      new DaemonThreadFactory("dd-trace-monitor");
-  public static final DaemonThreadFactory TRACE_PROCESSOR =
-      new DaemonThreadFactory("dd-trace-processor");
-  public static final DaemonThreadFactory TASK_SCHEDULER =
-      new DaemonThreadFactory("dd-task-scheduler");
+  public static final AgentThreadFactory TRACE_MONITOR = new AgentThreadFactory("dd-trace-monitor");
+  public static final AgentThreadFactory TRACE_PROCESSOR =
+      new AgentThreadFactory("dd-trace-processor");
+  public static final AgentThreadFactory TASK_SCHEDULER =
+      new AgentThreadFactory("dd-task-scheduler");
 
   private final String threadName;
   private final Runnable initializer;
 
   /**
-   * Constructs a new {@code DaemonThreadFactory} with a null ContextClassLoader.
+   * Constructs a new {@code AgentThreadFactory} with a null ContextClassLoader.
    *
    * @param threadName used to prefix all thread names.
    */
-  public DaemonThreadFactory(final String threadName) {
+  public AgentThreadFactory(final String threadName) {
     this(threadName, null);
   }
 
   /**
-   * Constructs a new {@code DaemonThreadFactory} with a null ContextClassLoader.
+   * Constructs a new {@code AgentThreadFactory} with a null ContextClassLoader.
    *
    * @param threadName used to prefix all thread names.
    * @param initializer initializer runnable that will be executed for all created threads.
    * @note The initializer runnable can be executed by multiple threads if more than one thread is
    *     created by the factory, and should take that into account to avoid race conditions.
    */
-  private DaemonThreadFactory(final String threadName, Runnable initializer) {
+  private AgentThreadFactory(final String threadName, Runnable initializer) {
     this.threadName = threadName;
     this.initializer = initializer;
   }
 
   /**
-   * Constructs a new {@code DaemonThreadFactory} with a null ContextClassLoader.
+   * Constructs a new {@code AgentThreadFactory} with a null ContextClassLoader.
    *
    * @param initializer initializer runnable that will be executed for all created threads.
    * @note The initializer runnable can be executed by multiple threads if more than one thread is
    *     created by the factory, and should take that into account to avoid race conditions.
    */
   public ThreadFactory withInitializer(Runnable initializer) {
-    return new DaemonThreadFactory(this.threadName, initializer);
+    return new AgentThreadFactory(this.threadName, initializer);
   }
 
   @Override
